pyproject.toml
[project]
name = "download-music"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.181.0",
    "google-auth-oauthlib>=1.2.2",
    "librosa>=0.11.0",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "shazamio>=0.8.1",
    "soundfile>=0.13.1",
    "spotipy>=2.25.1",
    "tensorflow>=2.20.0",
    "tensorflow-hub>=0.16.1",
    "yt-dlp[default]>=2025.9.23",
]

